Move editor UI layout code to dream-editor package
