+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+dream-editor = { path = "./dream-editor", version = "0.1.0" }
 instant = { version = "0.1", features = [ "stdweb" ] }
 cfg-if = "1"
 winit = "0.28"
@@ -43,3 +44,9 @@ web-sys = { version = "0.3", features = [
     "Window",
     "Element",
 ]}
+
+[workspace]
+members = [
+    "dream-editor",
+    "dream-resource",
+]
