+++ dream-editor/src/main.rs
@@ -1,3 +1,3 @@
 fn main() {
-    println!("Hello, world!");
+    println!("Hello world from Dream Editor!");
 }
